demo-site/requirements.txt
Flask
mysql-connector-python
gtts
